+++ railties/test/generators/migration_generator_test.rb
@@ -91,8 +91,9 @@ class MigrationGeneratorTest < Rails::Generators::TestCase
 
     assert_migration "db/migrate/#{migration}.rb" do |content|
       assert_method :change, content do |change|
-        assert_match(/remove_foreign_key :books, :authors/, change)
-        assert_no_match(/remove_foreign_key :books, :distributors/, change)
+        assert_match(/remove_reference :books, :author,.*\sforeign_key: true/, change)
+        assert_match(/remove_reference :books, :distributor/, change) # sanity check
+        assert_no_match(/remove_reference :books, :distributor,.*\sforeign_key: true/, change)
       end
     end
   end
@@ -189,8 +190,9 @@ class MigrationGeneratorTest < Rails::Generators::TestCase
 
     assert_migration "db/migrate/#{migration}.rb" do |content|
       assert_method :change, content do |change|
-        assert_match(/add_foreign_key :books, :authors/, change)
-        assert_no_match(/add_foreign_key :books, :distributors/, change)
+        assert_match(/add_reference :books, :author,.*\sforeign_key: true/, change)
+        assert_match(/add_reference :books, :distributor/, change) # sanity check
+        assert_no_match(/add_reference :books, :distributor,.*\sforeign_key: true/, change)
       end
     end
   end